requirements.txt
watchdog>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-mock>=3.0.0
pytest-cov>=4.0.0
